+++ netmanager/src/views/components/DataDisplay/DeviceView/DeviceOverview/index.js
@@ -1,0 +1,2 @@
+import DeviceOverview from "./DeviceOverview";
+export default DeviceOverview;
